[ch13_클래스의 상속]
- 추상화 기초 클래스
- 상속과 동적 메모리 대입
- 클래스 설계 복습

-----------------------------------

[알고리즘 수업 : 퀵정렬]

- 구현하는 중

+++ 02. 알고리즘 수업/ch06_Sort/ch06_Sort/Stopwatch.h
@@ -3,15 +3,15 @@
 class STOPWATCH_C
 {
 private:
-	float startHour;
-	float startMin;
-	float startSec;
-	float startMs;
+	int startHour;
+	int startMin;
+	int startSec;
+	int startMs;
 
-	float endHour;
-	float endMin;
-	float endSec;
-	float endMs;
+	int endHour;
+	int endMin;
+	int endSec;
+	int endMs;
 
 
 public:
